+++ src/app/agency-details/agency-details.component.ts
@@ -1,6 +1,8 @@
 import { Component } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
 import { MatTableModule } from '@angular/material/table';
+import { AgencyDetails } from '../types/gtfs.interface';
+import { AGENCY_DATA } from '../services/gtfs.service';
 
 @Component({
   selector: 'agency-details',
@@ -36,27 +38,3 @@ export class AgencyDetailsComponent {
  * Interface for the agency details
  * @Reference: https://gtfs.org/documentation/schedule/reference/#agencytxt
  */
-export interface AgencyDetails {
-  agencyId: string;
-  agencyName: string;
-  agencyUrl: string;
-  agencyTimezone: string;
-  agencyLang?: string;
-  agencyPhone?: string;
-  agencyFareUrl?: string;
-  agencyEmail?: string;
-}
-
-const AGENCY_DATA: AgencyDetails[] = [
-  {
-    agencyId: 'BST',
-    agencyName: 'Batik Solo Trans',
-    agencyUrl:
-      'https://pariwisatasolo.surakarta.go.id/wp-content/uploads/2019/01/Transit-Map-Solo.pdf',
-    agencyTimezone: 'Asia/Jakarta',
-    agencyLang: 'en',
-    agencyPhone: '',
-    agencyFareUrl: 'https://pariwisatasolo.surakarta.go.id/',
-    agencyEmail: '',
-  },
-];
