+++ src/app/gtfs-builder/gtfs-builder.component.html
@@ -30,6 +30,4 @@
         </mat-card-content>
     </mat-card>
 
-
-
 </div>
